Eradicated libpng cuz it didn't build + merged voidsprite/main

+++ freesprite/EditorBrushPicker.cpp
@@ -4,7 +4,7 @@
 
 bool EditorBrushPicker::isMouseIn(XY thisPositionOnScreen, XY mousePos)
 {
-	return pointInBox(mousePos, SDL_Rect{ thisPositionOnScreen.x, thisPositionOnScreen.y, wxWidth, wxHeight })
+    return pointInBox(mousePos, SDL_Rect{ thisPositionOnScreen.x, thisPositionOnScreen.y, wxWidth, wxHeight })
         || (patternsMenuOpen && patternMenuWidgets.mouseInAny(thisPositionOnScreen, mousePos));
 }
 
@@ -15,8 +15,8 @@ void EditorBrushPicker::render(XY position)
     SDL_Rect r = SDL_Rect{ position.x, position.y, wxWidth, wxHeight };
     //SDL_SetRenderDrawColor(g_rd, 0x30, 0x30, 0x30, focused ? 0x80 : 0x30);
     //SDL_RenderFillRect(g_rd, &r);
-    SDL_Color colorBG1 = { 0x30, 0x30, 0x30, focused ? 0xa0 : 0x90 };
-    SDL_Color colorBG2 = { 0x10, 0x10, 0x10, focused ? 0xa0 : 0x90 };
+    SDL_Color colorBG1 = { 0x30, 0x30, 0x30, uint8_t(focused ? 0xa0 : 0x90) };
+    SDL_Color colorBG2 = { 0x10, 0x10, 0x10, uint8_t(focused ? 0xa0 : 0x90) };
     renderGradient(r, sdlcolorToUint32(colorBG2), sdlcolorToUint32(colorBG1), sdlcolorToUint32(colorBG1), sdlcolorToUint32(colorBG1));
     if (focused) {
         SDL_SetRenderDrawColor(g_rd, 0xff, 0xff, 0xff, 255);
@@ -70,10 +70,10 @@ void EditorBrushPicker::eventButtonPressed(int evt_id)
         caller->currentPattern = g_patterns[evt_id - 60];
         updateActivePatternButton(caller->currentPattern);
     }
-	else if (evt_id >= 20) {
-		caller->currentBrush = g_brushes[evt_id-20];
+    else if (evt_id >= 20) {
+        caller->currentBrush = g_brushes[evt_id-20];
         updateActiveBrushButton(evt_id - 20);
-	}
+    }
 }
 
 void EditorBrushPicker::updateActiveBrushButton(int id)
